fix(jira-dashboard): isolate new frontend entrypoint
