+++ .changeset/fuzzy-ravens-wave.md
@@ -2,4 +2,4 @@
 '@axis-backstage/plugin-jira-dashboard': minor
 ---
 
-Add a configurable Jira user issues homepage widget for the new frontend system.
+Add a configurable Jira user issues homepage widget for the new frontend system and isolate the alpha entrypoint from legacy frontend APIs.
