+++ plugins/jira-dashboard/package.json
@@ -45,7 +45,6 @@
   "dependencies": {
     "@axis-backstage/plugin-jira-dashboard-common": "workspace:^",
     "@backstage/catalog-model": "^1.9.0",
-    "@backstage/core-compat-api": "^0.5.13",
     "@backstage/core-components": "^0.18.12",
     "@backstage/core-plugin-api": "^1.12.8",
     "@backstage/errors": "^1.3.1",
