+++ plugins/jira-dashboard/src/alpha.ts
@@ -1,11 +1,10 @@
-import { convertLegacyRouteRefs } from '@backstage/core-compat-api';
 import { createFrontendPlugin } from '@backstage/frontend-plugin-api';
 import {
   entityJiraContent,
   jiraApi,
   jiraUserIssuesWidget,
 } from './alpha/index';
-import { rootRouteRef } from './routes';
+import { rootRouteRef } from './alpha/routes';
 
 /**
  * Frontend plugin that fetches and displays Jira issues for an entity
@@ -16,7 +15,7 @@ import { rootRouteRef } from './routes';
 export default createFrontendPlugin({
   pluginId: 'jira-dashboard',
   extensions: [entityJiraContent, jiraApi, jiraUserIssuesWidget],
-  routes: convertLegacyRouteRefs({
+  routes: {
     rootRouteRef,
-  }),
+  },
 });
