+++ plugins/jira-dashboard/src/alpha/routes.ts
@@ -1,0 +1,4 @@
+import { createRouteRef } from '@backstage/frontend-plugin-api';
+
+/** @alpha */
+export const rootRouteRef = createRouteRef();
