+++ plugins/jira-dashboard/src/api/JiraDashboardApi.tsx
@@ -1,4 +1,4 @@
-import { createApiRef } from '@backstage/core-plugin-api';
+import { createApiRef, type ApiRef } from '@backstage/frontend-plugin-api';
 import {
   Issue,
   JiraResponse,
@@ -9,9 +9,11 @@ import {
  *
  * @public
  */
-export const jiraDashboardApiRef = createApiRef<JiraDashboardApi>({
-  id: 'plugin.jira-dashboard',
-});
+export const jiraDashboardApiRef: ApiRef<JiraDashboardApi> =
+  createApiRef<JiraDashboardApi>().with({
+    id: 'plugin.jira-dashboard',
+    pluginId: 'jira-dashboard',
+  });
 
 /**
  * The Jira dashboard API.
