+++ plugins/jira-dashboard/src/components/JiraDashboardContent/JiraDashboardContent.tsx
@@ -6,7 +6,7 @@ import {
   TabbedCard,
   CardTab,
 } from '@backstage/core-components';
-import { useApi } from '@backstage/core-plugin-api';
+import { useApi } from '@backstage/frontend-plugin-api';
 import { useEntity } from '@backstage/plugin-catalog-react';
 import { stringifyEntityRef } from '@backstage/catalog-model';
 import { jiraDashboardApiRef } from '../../api';
