+++ plugins/jira-dashboard/src/plugin.ts
@@ -7,24 +7,8 @@ import {
   fetchApiRef,
 } from '@backstage/core-plugin-api';
 import { rootRouteRef } from './routes';
-import { Entity } from '@backstage/catalog-model';
-import { PROJECT_KEY_NAME } from '@axis-backstage/plugin-jira-dashboard-common';
 import { jiraDashboardApiRef, JiraDashboardClient } from './api';
-
-/**
- * Checks if the entity has a jira.com project-key annotation.
- * @public
- * @param entity - The entity to check for the jira.com project-key annotation.
- */
-export const isJiraDashboardAvailable = (
-  entity: Entity,
-  annotationPrefix?: string,
-) =>
-  Boolean(
-    entity.metadata.annotations?.[
-      `${annotationPrefix ?? 'jira.com'}/${PROJECT_KEY_NAME}`
-    ],
-  );
+export { isJiraDashboardAvailable } from './availability';
 
 /**
  * Plugin that provides the Jira Dashboard api
